Conclusão pessoas por estado

+++ Banco-De-Sangue/src/main/java/application/Banco/De/Sangue/dto/ResultadosDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ResultadosDTO {
 
-    private int quantidadePorEstado;
+    private Long quantidadePorEstado;
     private String estado;
 }
